generic API code Added

+++ mvvmProduct/mvvmProduct/Screens/Product/Model/Product.swift
@@ -9,7 +9,7 @@ import Foundation
 struct Product : Decodable {
     let id : Int
     let title : String
-    let price : Float
+    let price : Double
     let description : String
     let category : String
     let image : String
@@ -17,6 +17,6 @@ struct Product : Decodable {
 }
 
 struct Rate : Decodable {
-    let rate : Float
+    let rate : Double
     let count : Int
 }
